create a EventStream GenServer
